fix a bug in convert_setup_tool(), aka G10
Tools have both a (constant) tool number and a (possibly variable)
pocket number.  This commit fixes a bug where these two were confused
with each other.

The result of the bug is that running G10 on a tool whose tool-number
differs form its pocket-number, while that tool is in the spindle (as
is done during Axis tool-table touch-off), would fail to update the
active TLO.