optimize and simplify find_tool_pocket()
This function searches the tool table and returns the slot/pocket
containing the tool with a given tool number.  It used to do this in a
somewhat silly way, this is cleaner (and incidentally a little quicker).

This changes the behavior of the function, it now returns the *first*
matching pocket instead of the *last* matching pocket, but I think a
particular tool number should be in at most one pocket, so this should
be ok.