add sim tests for the "toolno != pocket number" bug
These tests run G10 L10 and verify the behavior of linuxcnc using numbered
parameters for introspection.

There's currently a bug that results in incorrect TLO when using
nonrandom-tc and the tool number != pocket number.  (random-tc does not
have this bug.)  The fix is in the next few commits after this one.