Interp::convert_setup_tool(): avoid a spurious tool copy
When updating the tool table using G10, on a random tool changer, if the
tool being modified is currently loaded, then a spurious tool info copy
would be performed.

This commit checks if we're on a random tc machine, and if so skips the
tool copy.  (Also adds a comment describing the situation.)