sim: detect use of x87 instructions in nofp thread
This works by loading a value into every entry of the x87 floating-point
stack, setting a SIGFPE signal handler, and unmasking the invalid
operation exception.  Later, when an FP sequence is encountered, its
first instruction will be to load a value onto the stack, causing an
invalid operation exception, which triggers SIGFPE.  The SIGFPE handler
prints a backtrace and exits rtapi_app.

This is imperfect for several reasons:
 . the signal handler is global to all threads (but no other thread
   has a use of SIGFPE so this probably doesn't matter)
 . use of mmx/sse is not detected
 . it only works on x86