Refactored blending completely to abstract main calculations
Blend arc calculations before were a mess of local variables, and most
of the useful calculations were buried in the tpCreateBlendArc function.
This refactor pushes much of the actual math into the blendmath file.
Critically, this change also isolates the blend "parameters" from the
geometry, so that the same blend calculations can be reused for line-arc
blends, multi-axis blends, etc.