tp: fixed TP velocity target behavior
After fixing the last issue with tpGetRealTargetVelocity and
tpGetMaxTargetVelocity functions, the target velocity was artificially
limited to the cartesian vLimit, even for a rotary move. Now the target
velocity is no longer limited for this case. Furthermore, the
pre-calculated target velocity for spindle-synchronized motion is now
preserved (instead of being overridden by the cartesian limit).