tp: Overhaul of tangency check before creating blend arc.
Sam's testing with non-uniform axes revealed a weakness in the tangency
checking algorithm. Previously, we checked for tangency by computing the
angle between two segment's tangent lines, then estimating the corner
acceleration based on a simple heuristic. However, this method didn't
consider the acceleration direction, so we had to use a very
conservative maximum acceleration.

This new method evaluates how close to tangent two segments like so:
1) for the prev segment, compute acceleration required to completely decelerate to a stop in
one timestep
2) compute instantaneous acceleration to accelerate from rest for the
next segment.
3) Subtracting these large acceleration vectors estimates the "kink"
acceleration right at the intersection point.

From this kink accel vector, we element-wise divide by the allowed axis
acceleration so that it's normalized for each axis. If the largest
acceleration scale is under the threshold (currently 0.1), then we can
tolerate the "kink" in the path.