motion/interp: Massive overhaul to insert state tags at all levels.
This commit implements state tags at all levels. State tags contain a
snapshot of the interpreter state that rides along with a motion
command. This way, we can easily determine what the interpreter state
was at the time the motion was generated. This tag in principle can be
extended arbitrarily, though the size should be kept to a minimum.

Interp: generate state tags before calling a canon command that produces
a motion. Right now this is done by manually creating a tag before it's
needed.

Canon: Keeps a local copy of the state tag, which is updated by a
dedicated canon command. This tag is added to any outgoing motion, and
it's Interp's responsibility to keep it up to date. Ideally, canon
wouldn't have to deal with this at all, and the tag would be an argument
to canon commands (or packed into an appropriate structure).

Task/NML: State tags are included in the NML messages for motion
commands. HOWEVER, "update" methods are NOT implemented, because motion
messages are never actually transmitted via NML when Task / Motion are
on the same machine. In this case, the NML message is just a holding
tank for motion data, which is eventually copied into motion shared
memory.

Motion / TP: The state tag is included as an additional argument, but is
otherwise unused currently. Pretty much all that happens here is that
the tag is copied into the TC_STRUCT, and the currently executing
segment's tag is copied into motion's status output.

Status: If motion is running, it reports a state tag through
EMC_TRAJ_STAT. Task unpacks this into the appropriate active_g/m_codes
in emcStatus, so that UI's see the active state corresponding to the
motion. If motion is not running, then the active state shown is copied
directly from interp, as before.