tp: fixed spindle atspeed overrun due to prev line consumption
Running the 51 Meander Ave. example with the fast config led atspeed
errors at two points. This was because blending was consuming the
segments with the at-speed flag. Now we have an explicit check for
at-speed in addition to parabolic blending that prevents this.

To be safe, this fix should probably be generalized to a tcCanConsume
function that inspects other properties like IO changes.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>