tp: fall back to near-tangent case if blend arc is too slow
In some cases, the computed blend arc has a low maximum velocity:
1) If we're blending between two circular arcs that are nearly tangent
2) If the resulting arc is very short (and we hit the 1 segment / cycle
rule)

If the un-blended segments are nearly tangent, then we can get better
performance sometimes by simply slowing down and taking the corner
as-is.

This patch adds the calculations to do this safely. First, during the
tangent check, it calculates what the max safe velocity is for the pair
of nearly-tangent segments. If later, a blend arc between segments has a
lower maximum velocity, then we cap the final velocity of the segment to
the "kink" velocity, and reduce the acceleration of the segments to
match.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>