tp: Fix for sample maxvel calculation
Due to a quirk of how previous line segments are consumed, the effective
length of a blend arc is in some cases longer than the arc length
itself. There is also a short line segment "tail" that is part of the
motion. The new calculation factors this in, slightly increasing the
maximum velocity of blend arcs between short lines.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>