tp: don't allow arc blends between synced motions (falls back to 2.6 behavior)
The cause of bug #438 is that arc blend velocity is capped at the
maximum requested feed (factoring in override). If this feed is low,
then a blend arc between segments will have an artificially slow feed
during the blend.

The quick solution here is to simply disallow arc blending when either
segment is spindle synchronized. Now, the synced motion behavior should
match 2.6.x.

Arc blending and spindle-sync motion is possible with some re-work, but
this is the quickest way to restore normal behavior

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>